+++ sdks/cpp/include/thetadx.h
@@ -319,6 +319,7 @@ typedef struct {
     double charm;
     double vomma;
     double veta;
+    double vera;
     double speed;
     double zomma;
     double color;